Move utils to inside component folder
Waiting to refine the utils usage a little better before
making it look available for other components to use.
