+++ packages/components/src/visually-hidden/utils.js
